refactor: restructure project into modular multi-handler layout
