+++ .idea/gradle.xml
@@ -10,6 +10,10 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app-main" />
+            <option value="$PROJECT_DIR$/handler-mt-efs" />
+            <option value="$PROJECT_DIR$/handler-mx-gmps" />
+            <option value="$PROJECT_DIR$/shared-core" />
           </set>
         </option>
       </GradleProjectSettings>
