+++ handler-mt-efs/build.gradle
@@ -1,5 +1,7 @@
 plugins {
     id 'java'
+    id 'application'
+    id 'com.github.johnrengelman.shadow' version '8.1.1'
 }
 
 group = 'com.ccb.daily.file.pipeline'
@@ -12,6 +14,11 @@ repositories {
 dependencies {
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
+    implementation project(':shared-core')
+}
+
+application {
+    mainClass = 'com.ccb.daily.file.pipeline.EFSMain'
 }
 
 test {
@@ -21,4 +28,9 @@ test {
 java {
     sourceCompatibility = JavaVersion.VERSION_1_8
     targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+shadowJar {
+    archiveBaseName.set('handler-mt-efs-all')
+    archiveClassifier.set('')
 }
