+++ handler-mx-gmps/src/main/java/com/ccb/daily/file/pipeline/MXGMPSMain.java
@@ -1,0 +1,24 @@
+package com.ccb.daily.file.pipeline;
+
+public class MXGMPSMain {
+    public static void main(String[] args) throws Exception {
+        String siradt;
+        String tmdt;
+
+        if (args.length == 0) {
+            siradt = DateUtil.today();
+            tmdt = DateUtil.yesterday();
+        } else if (args.length == 2) {
+            siradt = args[0];
+            tmdt = args[1];
+        } else {
+            System.out.println("Usage: java -jar program.jar SIRA_DATE<yyyymmdd> TM_DATE<yyyymmdd>");
+            return;
+        }
+
+        MXMessageHandler handler = new MXMessageHandler();
+        ReportDateContext context = new ReportDateContext(siradt, tmdt);
+
+        handler.handle(context);
+    }
+}
