+++ shared-core/build.gradle
